refactor: create custom hooks for postUrl and storing on local storage
