+++ src/app/blog/page.js
@@ -2,63 +2,29 @@
 // import Loading from "./loading";
 import "./Blog.css";
 import BlogFilter from "../components/BlogFilter";
-import { useEffect, useState } from "react";
 import BlogItem from "./BlogItem";
+
+import usePostsUrl from "../hooks/usePostUrl";
+import useFetchPosts from "../hooks/useFetchPosts";
 
 export default function BlogPage({ searchParams }) {
   const searchQuery = searchParams?.search ?? "";
   const sortOptions = searchParams?.sortBy ?? "";
-  const [sortByValue, orderValue] = sortOptions.split("-");
-  let url = "https://dummyjson.com/posts";
 
-  const [postsUrl, setPostUrl] = useState(url);
-  const [posts, setPosts] = useState([]);
+  const { postsUrl } = usePostsUrl(searchQuery, sortOptions);
+  const { posts, setPosts } = useFetchPosts(postsUrl);
+  // const posts = [];
 
+  console.log(posts);
+
+  // Handle Delete
   function handleDelete(id) {
     localStorage.setItem(
       "posts",
       JSON.stringify(posts.filter((post) => post.id !== id))
     );
-    setPosts(JSON.parse(localStorage.getItem("posts")));
+    setPosts((curPosts) => curPosts.filter((post) => post.id !== id));
   }
-
-  useEffect(() => {
-    if (searchQuery) {
-      url = `https://dummyjson.com/posts/search?q=${searchQuery}`;
-      if (sortOptions) {
-        url += `&sortBy=${sortByValue}&order=${orderValue}`;
-      }
-    } else if (sortOptions) {
-      url = `https://dummyjson.com/posts?sortBy=${sortByValue}&order=${orderValue}`;
-    }
-
-    setPostUrl(url);
-  }, [searchQuery, sortOptions, sortByValue, orderValue]);
-
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      const savedPostsUrl = localStorage.getItem("postsUrl");
-      const savedPosts = localStorage.getItem("posts");
-
-      if (savedPosts && savedPostsUrl === postsUrl) {
-        setPosts(JSON.parse(savedPosts));
-      } else {
-        const fetchData = async () => {
-          try {
-            const response = await fetch(postsUrl);
-            const postsData = await response.json();
-
-            localStorage.setItem("posts", JSON.stringify(postsData.posts));
-            setPosts(postsData.posts);
-          } catch (error) {
-            console.error("Failed to fetch data", error);
-          }
-        };
-
-        fetchData();
-      }
-    }
-  }, [postsUrl]);
 
   return (
     <section className="blog-wrapper">
